refactor: move inputs to one component
