--- front/src/pages/User/LearnPage/BlockElement/FillTheGap/Answer/index.js
@@ -1,3 +1,0 @@
-import Answer from './Answer';
-
-export default Answer;
