+++ front/src/pages/User/LearnPage/BlockElement/FillTheGap/FillTheGap.styled.js
@@ -1,0 +1,12 @@
+import { Button } from 'antd';
+import styled from 'styled-components';
+
+export const LessonButton = styled(Button).attrs({
+  size: 'large',
+})`
+  @media (max-width: 767px) {
+    margin-top: auto;
+  }
+  width: 100%;
+  margin-bottom: 1rem;
+`;
