+++ front/src/pages/User/LearnPage/BlockElement/FillTheGap/Inputs/index.js
@@ -1,0 +1,3 @@
+import Inputs from './Inputs';
+
+export default Inputs;
