--- front/src/pages/User/LearnPage/BlockElement/FillTheGap/Result/AnswerResult/index.js
@@ -1,3 +1,0 @@
-import AnswerResult from './AnswerResult';
-
-export default AnswerResult;
