+++ front/src/pages/User/LearnPage/BlockElement/FillTheGap/Result/Result.jsx
@@ -1,49 +1,60 @@
-import { useMemo } from 'react';
+import { Typography } from 'antd';
+import PropTypes from 'prop-types';
+import { useTranslation } from 'react-i18next';
+import { CheckCircleTwoTone, CloseCircleTwoTone } from '@ant-design/icons';
 
-import {
-  FillTheGapBlockAnswerType,
-  FillTheGapBlockDataType,
-} from '@sb-ui/pages/User/LearnPage/BlockElement/types';
-import { ChunkWrapper } from '@sb-ui/pages/User/LearnPage/LearnPage.styled';
+import Inputs from '@sb-ui/pages/User/LearnPage/BlockElement/FillTheGap/Inputs';
 
-import { verifyAnswers } from '../verifyAnswers';
+import { CORRECT_ALL, CORRECT_NONE, CORRECT_PARTIAL } from '../constants';
 
-import AnswerResult from './AnswerResult';
 import * as S from './Result.styled';
 
-const Result = ({ correctAnswer, data }) => {
-  const { text, answers } = data;
-  const { results } = correctAnswer;
-  const chunks = useMemo(() => text?.split('{{ # }}'), [text]);
+const { Text } = Typography;
 
-  const { correct, result } = verifyAnswers(results, answers);
+const Result = ({ correct, result, text }) => {
+  const { t } = useTranslation('user');
+
+  if (correct === CORRECT_ALL) {
+    return (
+      <S.AnswerWrapper>
+        <Text>{t('lesson.answer_result.correct')}</Text>
+        <CheckCircleTwoTone twoToneColor="#52c41a" />
+      </S.AnswerWrapper>
+    );
+  }
+
   return (
     <>
-      <ChunkWrapper>
-        <S.ChunkText>
-          {chunks.map((chunk, index) => {
-            if (index === chunks.length - 1) {
-              return <span>{chunk}</span>;
-            }
-            return (
-              <>
-                <span>{chunk}</span>
-                <S.Input value={answers[index]} />
-              </>
-            );
-          })}
-        </S.ChunkText>
-      </ChunkWrapper>
-      <ChunkWrapper>
-        <AnswerResult correct={correct} result={result} chunks={chunks} />
-      </ChunkWrapper>
+      {correct === CORRECT_PARTIAL && (
+        <S.AnswerWrapper>
+          <Text>{t('lesson.answer_result.partially_wrong')}</Text>
+          <S.WarningTwoTone twoToneColor="#FADB14" />
+        </S.AnswerWrapper>
+      )}
+      {correct === CORRECT_NONE && (
+        <S.AnswerWrapper>
+          <Text>{t('lesson.answer_result.wrong')}</Text>
+          <CloseCircleTwoTone twoToneColor="#F5222D" />
+        </S.AnswerWrapper>
+      )}
+      <S.ResultWrapper>
+        <Text italic>
+          <Inputs text={text} result={result} />
+        </Text>
+      </S.ResultWrapper>
     </>
   );
 };
 
 Result.propTypes = {
-  data: FillTheGapBlockDataType,
-  correctAnswer: FillTheGapBlockAnswerType,
+  correct: PropTypes.oneOf([CORRECT_ALL, CORRECT_NONE, CORRECT_PARTIAL]),
+  result: PropTypes.arrayOf(
+    PropTypes.shape({
+      correct: PropTypes.bool,
+      value: PropTypes.string,
+    }),
+  ),
+  text: PropTypes.string,
 };
 
 export default Result;
